Fix duplicate query IDs
It's possible to have duplicate query IDs if your CPU is fast enough.
This is fixed by maintaining a set of in-flight query IDs.

Furthermore, this also makes my local systemd-resolved service
appreciate this library as it keeps working after running stress tests.
Before this diff, it broke and I couldn't use Firefox any more without
restarting the systemd-resolved service.

The stress test was also executed with a local pdns-recursor.