Remove invalid includes
Because the headers must be installed in /usr/include/dnscpp before
compiling a test file, it may happen that a header file that did exist
previously, but not anymore, lingers in that directory. This makes
it somewhat treacherous to include files. It's probably best to remove
the /usr/include/dnscpp directory before doing a `make install`.