Allow using multiple UDP sockets
This is independent of the number of nameservers. So far I don't
see much improvement in outgoing kb/s in network traffic.

There's still a close() method but it closes all sockets. Closing
of sockets should instead be decided elsewhere and should be
granular per socket. We would need to keep track of things like
the number of inflight requests per socket for that.