Fixes project 38464
It can be the case that while we're connecting to the TCP socket the
Lookup might die in the meantime. The TCP socket kept a list of
"connectors" that it would inform when it was connected. This is now
a list of weak pointers to connectors to make sure that we're not
calling into dangling pointers.